First attempt at a port.
This required pulling in some types and utility functions from other
assorted files.  I have tried to leave them in the same order as seen
in their respective files, and also attempted to retain the same file
names for tracking where the definitions have come from in the source
implementation.

Signed-off-by: Daniel Farina <daniel@fdr.io>